granulator additons
granulator record-method now returns an option, where Some() returns the
input sample as a passthrough. The use-case is to record until buffer is
filled or if aborted prematurely. can be used to record shorter content
into buffer, but the size of the buffer unless completly filled is not
yet handled.

The None can be used to handle switching behavior between recording
and play
-back as seen in playground file