new designs on oscillators
redesign on wavetable, polytable and vector oscillators.
the main idea is to supply a borrowed array of a wavetable to a
play-function instead of letting the oscillator own it, or use
messy counted references that add overhead.

leaning into iterators and redesigning functions where looping over
data arrays of same length occurs, converting several arrays into a
single array of some struct