better benchmarks
redemption for the shared wavetable, it performs almost as fast as
the owned-variant, which does not have to access the table as a
reference, and it performs faster than the Arc<RwLock<Vec<f32>>>-
version. This makes it easier to handle sharing tables across
multiple voices of a polyphonic synth.

redemption for the granulator leveraging iterators and with cleaner
play-function. using unsafe to do an unchecked get of the next grain
-to-be-triggereds values, which may or may not be faster, but skips
the automatic bounds-check that rust does per default. The speed-
difference between the old granulator and new is almost negligible.