new envelope design + trigger class impl callbacks
envelope design based on break points. increments every tick, different
to the previous design that generated a huge wavetable based on
wavetables. unecessarily big lookup tables are now swapped for a smaller
array of breakpoints.

adds callbacks to trigger objects, being able to call a closure, for
example calling a number of objects clocked by a trigger generator.
triggering had previously been handled naively, where the trig generator
generated a f32 signal, with a check to see if the trig was valid.