accidental waveshape and wavetable rewrite
redesigned the implementation of the complex_sine function, to not use
that many special generics.

redesigned the shared::Wavetable struct, to not store a copy of
samplerate, but only a copy of the reciprocal: 1.0 / samplerate.
since samplerate in its original form is not used, we do not need to
save it.
- was previously used as a filter to disable sound generation if freq is
above nyqvist, but it is an unnecessary branch.